TASK: Adjust tests that exactly once delivery is not possible for external projections
see 175ab4cefe37a6d6565801fe38856f8f6ce4fce8
